# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Sample CRUD",
    "description": "Sample CRUD API for a To-do List",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Base localhost for server.js file"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Returns list of tasks.",
        "description": ".",
        "responses": {
          "200": {
            "description": "Returns array of items consisting of id, title, and done (status).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "done": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Creates a new task.",
        "requestBody": {
          "description": "Task to create",
          "required": [
            "title"
          ],
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created task"
          }
        }
      }
    },
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Returns current server status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Returns a task with specified id number.",
        "description": ".",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "id of task to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns single object consisting of id, title, and done (status).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "done": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update existing task.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "id of task to get"
          }
        ],
        "requestBody": {
          "description": "Task to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "done": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task"
          },
          "404": {
            "description": "Task not found"
          },
          "400": {
            "description": "Invalid body"
          }
        }
      },
      "delete": {
        "summary": "Delete a task with specified id.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "id of task to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  }
}